requirements.txt
# Conda dependencies converted for pip where possible
libgcc_mutex==0.1
blas==1.0
ca-certificates==2020.1.1
certifi==2020.4.5.1
cffi==1.14.0
cmake==3.14.0
cudatoolkit==10.0.130
cycler==0.10.0
dbus==1.13.12
expat==2.2.6
faiss-gpu==1.6.3
freetype==2.9.1
glib==2.63.1
graphite2==1.3.13
gst-plugins-base==1.14.0
gstreamer==1.14.0
h5py==2.8.0
icu==58.2
imageio==2.8.0
intel-openmp==2020.0
joblib==0.14.1
jpeg==9b
kiwisolver==1.1.0
krb5==1.17.1
ld_impl_linux-64==2.33.1
libcurl==7.69.1
libedit==3.1.20181209
libffi==3.2.1
libgcc-ng==9.1.0
libgfortran-ng==7.3.0
libprotobuf==3.11.4
libssh2==1.9.0
libstdcxx-ng==9.1.0
libtiff==4.1.0
libxcb==1.13
libxml2==2.9.9
matplotlib==3.1.3
matplotlib-base==3.1.3
mkl==2020.0
mkl-service==2.3.0
mkl_fft==1.0.15
mkl_random==1.1.0
ncurses==6.2
ninja==1.9.0
numpy==1.18.1
numpy-base==1.18.1
olefile==0.46
opencv==3.4.2
openssl==1.1.1g
pcre==8.43
pillow==7.0.0
pip==20.0.2
pixman==0.38.0
protobuf==3.11.4
py-opencv==3.4.2
pycparser==2.20
pyparsing==2.4.6
pyqt==5.9.2
python==3.7.7
python-dateutil==2.8.1
pytorch==1.4.0
pyyaml==5.3.1
qt==5.9.7
readline==8.0
rhash==1.3.8
scikit-learn==0.22.1
setuptools==46.1.3
sip==4.19.8
six==1.14.0
sqlite==3.31.1
swig==3.0.12
termcolor==1.1.0
tk==8.6.8
torchvision==0.5.0
tornado==6.0.4
typing==3.6.4
wheel==0.34.2
xz==5.2.4
yaml==0.1.7
zlib==1.2.11
zstd==1.3.7

# Pip dependencies
blis==0.4.1
catalogue==1.0.0
chardet==3.0.4
cymem==2.0.3
idna==2.9
importlib-metadata==1.6.0
murmurhash==1.0.2
plac==1.1.3
preshed==3.0.2
requests==2.23.0
spacy==2.2.4
srsly==1.0.2
thinc==7.4.0
tqdm==4.45.0
urllib3==1.25.8
wasabi==0.6.0
zipp==3.1.0
easydict==1.9
tensorboardx==2.1
